+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[3.5.0] - 2021-11-19
+### Added
+- Azure DevOps scenario content: scenario updates.
+
+[3.4.1] - 2021-11-19
+### Changed
+- Exit status code has been changed to 0 (success) when requested help or version.
+
 ## [3.4.0] - 2021-11-10
 ### Added
 - Azure DevOps scenario content: area path configuration.